head to my base

+++ src/main/java/com/javachina/ext/Funcs.java
@@ -1,11 +1,10 @@
 package com.javachina.ext;
 
+import com.blade.kit.DateKit;
+import com.blade.kit.StringKit;
 import com.javachina.Constant;
 import com.javachina.ImageTypes;
-import com.javachina.kit.DateKit;
 import com.javachina.kit.Utils;
-
-import blade.kit.StringKit;
 
 public class Funcs {
 
@@ -47,10 +46,10 @@ public class Funcs {
 	
 	/**
 	 * 格式化日期
-	 * @param date
+	 * @param unixTime
 	 * @return
 	 */
-	public static String fmtdate(Long unixTime) {
+	public static String fmtdate(Integer unixTime) {
 		if(null != unixTime){
 			return DateKit.formatDateByUnixTime(unixTime, "yyyy-MM-dd");
 		}
@@ -59,11 +58,11 @@ public class Funcs {
 	
 	/**
 	 * 格式化日期
-	 * @param date
+	 * @param unixTime
 	 * @param patten
 	 * @return
 	 */
-	public static String fmtdate(Long unixTime, String patten) {
+	public static String fmtdate(Integer unixTime, String patten) {
 		if(null != unixTime && StringKit.isNotBlank(patten)){
 			return DateKit.formatDateByUnixTime(unixTime, patten);
 		}
@@ -71,7 +70,7 @@ public class Funcs {
 	}
 	
 	public static String today(String patten){
-		return fmtdate((long) DateKit.getCurrentUnixTime(), patten);
+		return fmtdate(DateKit.getCurrentUnixTime(), patten);
 	}
 	
 	/**
@@ -93,10 +92,10 @@ public class Funcs {
 	/**
 	  * 显示时间，如果与当前时间差别小于一天，则自动用**秒(分，小时)前，如果大于一天则用format规定的格式显示
 	  * 
-	  * @param ctime时间
+	  * @param ctime 时间
 	  * @return
 	  */
-	public static String timespan(Long ctime) {
+	public static String timespan(Integer ctime) {
 		String r = "";
 		if (ctime == null)
 			return r;
